move GUIDialogFavourites to favourites module
