+++ xbmc/favourites/CMakeLists.txt
@@ -1,7 +1,9 @@
 set(SOURCES ContextMenus.cpp
+            GUIDialogFavourites.cpp
             FavouritesService.cpp)
 
 set(HEADERS ContextMenus.h
+            GUIDialogFavourites.h
             FavouritesService)
 
 core_add_library(favourites)
