+++ xbmc/favourites/GUIDialogFavourites.h
